requirements.txt
torch>=2.5.0
lightning>=2.5.0
transformers>=5.0.0
peft>=0.17.0
accelerate>=0.20.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0
biopython>=1.79
anndata>=0.9.0
scanpy>=1.9.0
h5py>=3.8.0
sentencepiece>=0.1.99
tokenizers>=0.13.0
safetensors>=0.3.0
optuna>=3.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
colorlog>=6.7.0
wandb>=0.15.0
pyyaml>=6.0
huggingface-hub>=0.16.0
torchmetrics>=1.0.0
click>=8.0.0
joblib>=1.2.0
datasets>=4.5.0
fsspec>=2023.1.0
requests>=2.28.0
typing-extensions>=4.5.0
statsmodels>=0.14.0
patsy>=0.5.3
umap-learn>=0.5.3
pynndescent>=0.5.10
zarr>=2.14.0
pyarrow>=12.0.0
